# api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StockNew API",
    "description": "Local async API surface over the StockNew quant platform.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/scans": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Trigger Scan",
        "description": "Run a synchronous scan and persist the result.\n\nSynchronous from the caller's perspective \u2014 blocks until the scan\ncompletes. Heavy compute runs in a worker thread so the event loop stays\nresponsive. Phase 1.7 adds /api/stream/scan-progress for live updates.",
        "operationId": "trigger_scan_api_scans_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "List Scans",
        "description": "Most recent scan runs, newest first.",
        "operationId": "list_scans_api_scans_get",
        "parameters": [
          {
            "name": "strategy",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strategy"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "exclusiveMinimum": 0,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScanSummary"
                  },
                  "title": "Response List Scans Api Scans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scans/{run_id}": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Get Scan",
        "operationId": "get_scan_api_scans__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/recommendations": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "List Recommendations",
        "description": "Most recent paper-trading recommendations, newest first.",
        "operationId": "list_recommendations_api_recommendations_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ticker"
            }
          },
          {
            "name": "strategy",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strategy"
            }
          },
          {
            "name": "submitted_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Submitted Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "exclusiveMinimum": 0,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaperRecommendationItem"
                  },
                  "title": "Response List Recommendations Api Recommendations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recommendations/{rec_id}": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Get Recommendation",
        "operationId": "get_recommendation_api_recommendations__rec_id__get",
        "parameters": [
          {
            "name": "rec_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rec Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperRecommendationItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PaperRecommendationItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "scan_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Scan Timestamp"
          },
          "composite_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Composite Score"
          },
          "action": {
            "type": "string",
            "enum": [
              "STRONG BUY",
              "BUY",
              "HOLD",
              "SELL",
              "STRONG SELL"
            ],
            "title": "Action"
          },
          "sub_scores": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Sub Scores"
          },
          "entry_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Price"
          },
          "stop_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss"
          },
          "take_profit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "earnings_in_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earnings In Days"
          },
          "submitted": {
            "type": "boolean",
            "title": "Submitted",
            "default": false
          },
          "skip_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skip Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ticker",
          "strategy",
          "scan_timestamp",
          "composite_score",
          "action"
        ],
        "title": "PaperRecommendationItem"
      },
...
      "ScanRequest": {
        "properties": {
          "strategy": {
            "type": "string",
            "title": "Strategy",
            "default": "swing_trading"
          },
          "budget": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget"
          },
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "top": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 200.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top"
          },
          "fresh": {
            "type": "boolean",
            "title": "Fresh",
            "description": "Bypass cache, fetch live data",
            "default": false
          }
        },
        "type": "object",
        "title": "ScanRequest"
      },
      "ScanResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "scan_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Scan Timestamp"
          },
          "n_candidates": {
            "type": "integer",
            "title": "N Candidates"
          },
          "n_results": {
            "type": "integer",
            "title": "N Results"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ScanResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "strategy",
          "scan_timestamp",
          "n_candidates",
          "n_results",
          "results"
        ],
        "title": "ScanResponse"
      },
      "ScanResultItem": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "action": {
            "type": "string",
            "enum": [
              "STRONG BUY",
              "BUY",
              "HOLD",
              "SELL",
              "STRONG SELL"
            ],
            "title": "Action"
          },
          "composite_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Composite Score"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "sub_scores": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Sub Scores"
          },
          "reasoning": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasoning"
          },
          "bullish_signals": {
            "type": "integer",
            "title": "Bullish Signals",
            "default": 0
          },
          "bearish_signals": {
            "type": "integer",
            "title": "Bearish Signals",
            "default": 0
          },
          "breakdown": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Breakdown"
          },
          "risk_management": {
            "additionalProperties": true,
            "type": "object",
            "title": "Risk Management"
          },
          "sector": {
            "type": "string",
            "title": "Sector",
            "default": "Unknown"
          },
          "industry": {
            "type": "string",
            "title": "Industry",
            "default": "Unknown"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "market_cap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "action",
          "composite_score",
          "confidence"
        ],
        "title": "ScanResultItem",
        "description": "One recommendation row in a scan response. Permissive shape \u2014 accepts\nthe existing recommender dict; web layer narrows what it renders."
      },
      "ScanSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "scan_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Scan Timestamp"
          },
          "n_candidates": {
            "type": "integer",
            "title": "N Candidates"
          },
          "top_ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Ticker"
          },
          "top_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Score"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "strategy",
          "scan_timestamp",
          "n_candidates"
        ],
        "title": "ScanSummary",
        "description": "Lighter representation for the GET /api/scans list view."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}